Use Integer type to consistency with AudioMessageContent.

+++ line-bot-model/src/main/java/com/linecorp/bot/model/event/message/VideoMessageContent.java
@@ -34,6 +34,9 @@ public class VideoMessageContent implements MessageContent {
     }
 
     String id;
-    Long duration;
+    /**
+     * Length of video file (milliseconds)
+     */
+    Integer duration;
     ContentProvider contentProvider;
 }
